package iproute2 has problems building on recent BSD-derived platforms because some programs to be run on the host use malloc.h in an incompatible way.
package/iproute2/patches/004-darwin_fixes.patch has a fix for Darwin,
however the fix can be made to work on other BSD platforms
by changing the patch as below.

I think a similar approach should work for other BSD-derived platforms,
replacing the
	#ifndef __APPLE__
with
	#if !defined(__APPLE__) && !defined(__FreeBSD__) && !defined(...)

or perhaps even with just

	#ifdef __linux__

if the set of platforms where malloc.h is ok is small or restricted
to just linux

In any case, the patch is below. (#3869)
Signed-off-by: Luigi Rizzo <rizzo@iet.unipi.it> 


git-svn-id: svn://svn.openwrt.org/openwrt/trunk@12249 3c298f89-4303-0410-b956-a3cf2f4a3e73