package/u-boot-envtools: update to version 2011.06.
This updates uboot-envtools to the version which comes with U-Boot 2011.06.
The important novelity is that fw_setenv now supports batch processing, which
greatly improves sysupgrade of the ALL0258N as otherwise we would have to
overwrite uboot-env up to 8x in a single firmware upgrade.
Unfortunately, no newer version of the sources is available from Debian, so we
got to take it from the u-boot source which is kinda ugly...

Signed-off-by: Daniel Golle <dgolle@allnet.de>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@28566 3c298f89-4303-0410-b956-a3cf2f4a3e73