ar71xx: local mac support on the wndr3700
The wndr3700 at least has no eth0 mac address and usually leverages
the first wireless device's mac when in a bridged scenario. If,
however, you want to route, and not bridge the interfaces, you
need a unique mac address for it.

This patch sets the local bit on the mac address pulled from the
wireless chip and uses the resulting address for eth0.

Patch-by: Dave Taht <dave.taht@gmail.com>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@29675 3c298f89-4303-0410-b956-a3cf2f4a3e73