ar71xx: Buffalo WLAE-AG300N initial support
The patch set for Buffalo WLAE-AG300N initial support.

There is another patch for wireless led support that
is posted separately.

Note on serial console:
This unit has buffalo standard 4 pin console, but the unit may not power
on if some console apparatus is connected. This is probably due to some
electronic interaction between the unit's electronic power switch circuit
and the serial console apparatus. If this happens, it is required to power
on the unit without the console, then quickly (re-)attach it.

[juhosg: fix coding style]

Signed-off-by: Yoichi Shinoda <shinoda@jaist.ac.jp>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@32931 3c298f89-4303-0410-b956-a3cf2f4a3e73