package/uboot-envtools: always install the fw_setenv symlink
It is required by sysupgrade on some boards, and it is only
working if a suitable /etc/fw_env.config is present on the
target board. If an user creates such configuration he/she
should know that it is a dangerous tool. The devices can
be bricked in several other ways anyway.

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@33265 3c298f89-4303-0410-b956-a3cf2f4a3e73