broadcom-diag: support for the WNDR3700v3
I recently picked up a WNDR3700 to put OpenWRT on, and only after tearing into the box did I find it 
was one of the v3 boards, with poor OpenWRT support. This patch should add the board detection and
LED/button control to the broadcom-diag module, and should generate a netgear .chk image that the
bootloader and stock firmware will accept.

The changes to the broadcom-diag module are more than a few lines because the WNDR3700v3 is driving
its LEDs through an HC164 8-bit shift register.

Signed-off-by: Owen Kirby <osk@exegin.com>


git-svn-id: svn://svn.openwrt.org/openwrt/trunk@36482 3c298f89-4303-0410-b956-a3cf2f4a3e73