kernel: adm6996: add support for ADM6996L and GPIO interface
This patch makes it possible to use adm6996.c on first generation 
BCM47XX devices with ADM switches.
The GPIO bit banging protocol implementation was copied from the old 
switch driver and adapted to this driver and changed to the mainline 
kernel GPIO interface.
The ADM6996L is different from the ADM6996M which is supported, for 
both specs are available in the Internet.

This was tested on a WRT54GS version 1.0, thank you Dirk Neukirchen for 
the device.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>


git-svn-id: svn://svn.openwrt.org/openwrt/trunk@38698 3c298f89-4303-0410-b956-a3cf2f4a3e73