ixp4xx: Don't overwrite memory in wg302v1_fixup()
wg302v1_fixup() looks for the ATAG_CMDLINE atag, it doesn't find it and
so it takes the last atag in the list and overwrites non allocated
memory.

The side effects are corrupted vital data and a kernel that doesn't
boot. More details here https://dev.openwrt.org/ticket/18356

The fix appends the fixup to the command line and updates the atag only
when it finds it.

Signed-off-by: Gianluca Anzolin <gianluca@sottospazio.it>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@43320 3c298f89-4303-0410-b956-a3cf2f4a3e73