ramips: improve and fix Memory 2 Move support
This patch is a follow up for my previous patch:
"ramips: add support for Intenso Memory 2 Move USB 3.0".

It fixes a couple of errors in the DTS (one of which broke
the gpio-buttons). The kmod-leds-gpio dependency has been
dropped as it is already part of the ramips target.

Furthermore the ramdisk/uImage image is generated by default
for the rt3050 subtarget. This image is needed to flash
OpenWrt for the first time onto the device via TFTP.

Signed-off-by: Christian Lamparter <chunkeey@googlemail.com>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@44072 3c298f89-4303-0410-b956-a3cf2f4a3e73