lantiq: correct value for fpi clock on ar9
This adds a patch to target/linux/lantiq/patches-3.14
fixing a bug clock code on ar9. The current version returns
the wrong value for the fpi clock frequency in some
cases.

See discussion for further details:
https://lists.openwrt.org/pipermail/openwrt-devel/2015-January/030688.html

I'm not sure about the patch naming and numbering convention.
Do please let me know it this is not OK.

Many thanks,

Ben Mulvihill

Signed-off-by: Ben Mulvihill <ben.mulvihill@gmail.com>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@44083 3c298f89-4303-0410-b956-a3cf2f4a3e73