sunxi: create KDIR_TMP, build boot.img in it
It's nice to have a distinction between files which are built during the
normal OpenWrt build process and ones that are created by the ImageBuilder
(like it already exists on some other targets).

This also gives boot.img a profile-specific prefix as it contains profile-
specific data.

KDIR_TMP does not need to be specified since include/image.mk already sets
it.

Signed-off-by: Matthias Schiffer <mschiffer@universe-factory.net>
Signed-off-by: Zoltan HERPAI <wigyori@uid0.hu>



git-svn-id: svn://svn.openwrt.org/openwrt/trunk@46607 3c298f89-4303-0410-b956-a3cf2f4a3e73