ramips: Enable RTC support for PBR-M1
The PBR-M1 and other upcoming MT7621 boards have RTC chips on them. The
PBR-M1 also selects the kmod-rtc-pcf8563 by default. But the module itself
will not be build because CONFIG_RTC_CLASS is currently not enabled for its
kernel.

Enabling this option should fix the problem of the missing rtc device on
these boards.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@46857 3c298f89-4303-0410-b956-a3cf2f4a3e73