ar71xx: serial: core: add support for boot console with arbitrary baud rates
The Arduino Yun uses a baud rate of 250000 by default. The serial is
going over the Atmel ATmega and is used to connect to this chip.
Without this patch Linux wants to switch the console to 230400 Baud.

With this patch Linux will use the configured baud rate and not some
standard one which is near by.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>


git-svn-id: svn://svn.openwrt.org/openwrt/trunk@47450 3c298f89-4303-0410-b956-a3cf2f4a3e73