ath10k-firmware: update firmware version for QCA988X (again)
Updates the ath10-firmware release to current git and updates the
QCA988X firmware in use to version 10.2.4.70.12-2 which appears to
actually fix the stability problems when using a Nexus 5X phone as
per issue #20854. Also updates the Makefile for a change in the
location of the board.bin file for QCA988X in the source repository.

Signed-off-by: Robert Hancock <hancockrwd@gmail.com>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@47469 3c298f89-4303-0410-b956-a3cf2f4a3e73