ar71xx: fix wndr3700_board_detect for some NETGEAR WNDR3700v2 (again)
When fixing the model string for WNDR3700v2 which contain a model string
followed by 0xff in r46455, the match for other versions of the WNDR3700v2
which just contain lots of 0xff broke (as the 0xff $model is checked for
is stripped off).

Fix by stripping off non-printable characters only for the actual output
string, but not for the internal matching.

Signed-off-by: Matthias Schiffer <mschiffer@universe-factory.net>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@47538 3c298f89-4303-0410-b956-a3cf2f4a3e73