lantiq: Configure disabled WLAN bands for TDW89X0.
The TDW8970 has a AR9381, which is the bgn 3x3:3 variant of the AR938x family.

The TDW8980 has a AR9287, which is the bgn 2x2:2 variant of the AR928x family.

This means that the chip for both routers is 2.4 GHz only.

Anyway, the manufacturer didn't disable the 5 GHz band in the EEPROM partition
(at least on my TDW8970).

So this patch disables the 5 GHz band.

Signed-off-by: Vittorio Gambaletta <openwrt@vittgam.net>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@47912 3c298f89-4303-0410-b956-a3cf2f4a3e73